Refactor code for preserving music data
