+++ data/music/music.json
@@ -1,1 +1,1 @@
-[{"name":"For My Brothers","publish_date":1572739200,"archived":false,"on_spotify":true,"id":29608,"genre_id":"101"},{"name":"Asa","publish_date":1287014400,"archived":false,"on_spotify":true,"id":36087,"genre_id":"106"}]
+[{"name":"For My Brothers","publish_date":1572739200,"archived":false,"on_spotify":true,"id":29608,"genre_id":"101"},{"name":"hello","publish_date":1607731200,"archived":false,"on_spotify":false,"id":44331,"genre_id":"108"},{"name":"r","publish_date":1607731200,"archived":false,"on_spotify":false,"id":54525,"genre_id":"103"},{"name":"test","publish_date":1607731200,"archived":false,"on_spotify":false,"id":2088,"genre_id":"104"},{"name":"another test","publish_date":1607731200,"archived":false,"on_spotify":false,"id":10874,"genre_id":"109"},{"name":"8","publish_date":661478400,"archived":false,"on_spotify":false,"id":28510,"genre_id":"102"}]
