+++ main.rb
@@ -2,6 +2,7 @@ require_relative './app/app'
 
 def main
   app = App.new
+  app.welcome
   app.load_data
   app.console_start
 end
